+++ eva/src/components/dashBoardInside/dashConstructorScheme/dashConstructorSchemes.vue
@@ -567,12 +567,12 @@
       </div>
       <!--Settings-element-panel-->
       <div
-        class="dash-constructor-schemes__data-panel"
+        class="dash-constructor-schemes__data-panel pr-0"
         :class="{
           'dash-constructor-schemes__data-panel--active': dataPanel,
         }"
       >
-        <div class="row mb-3">
+        <div class="row">
           <div class="col-12">
             <div
               class="d-flex justify-content-right"
@@ -1470,7 +1470,7 @@ export default {
     top: 5px;
     bottom: 15px;
     background-color: var(--main_bg);
-    width: 250px;
+    width: 260px;
     padding: 10px;
     transition: all .2s ease;
     pointer-events: none;
@@ -1492,6 +1492,7 @@ export default {
     }
   }
   &__data-panel {
+    overflow: hidden;
     right: 0;
     width: 300px;
     transform: translateX(100%);