+++ eva/src/components/dashBoardInside/dashConstructorScheme/dashConstructorSchemesSettings.vue
@@ -2,9 +2,9 @@
   <div
     v-if="dataObject"
     ref="dataPanelItems"
-    class="dash-constructor-schemes__data-panel-wrapper"
+    class="dash-cs__data-panel-wrapper"
   >
-    <div class="dash-constructor-schemes__data-panel-item pb-4">
+    <div class="dash-cs__data-panel-item pb-4">
       <template v-if="dataType === 'data-type-0'">
         <!--data-type-0-->
         <div class="row justify-content-between align-center">
@@ -21,8 +21,28 @@
         <div
           v-for="(element, index) in dataObject.items"
           :key="`${dataObject.nodeId}-${index}`"
-          class="column dash-constructor-schemes__data-type-0"
+          class="column dash-cs__data-type-0"
         >
+          <v-btn
+            v-if="dataObject.items.length > 1"
+            ripple
+            small
+            height="20"
+            min-height="20"
+            width="20"
+            min-width="20"
+            class="dash-cs__close-btn px-0 py-0"
+            color="transparent"
+            @click="deleteLine(index, 'items')"
+          >
+            <v-icon
+              class="theme--dark px-0 py-0"
+              size="20"
+              :style="{ color: theme.$main_text }"
+            >
+              {{ closeIcon }}
+            </v-icon>
+          </v-btn>
           <div class="row align-center relative">
             <v-text-field
               v-if="textMode"
@@ -48,7 +68,7 @@
               <template v-slot:item="{ item, on }">
                 <v-list-item
                   ripple
-                  class="v-list-item--link dash-constructor-schemes__data-panel-select"
+                  class="v-list-item--link dash-cs__data-panel-select"
                   v-on="on"
                 >
                   <v-list-item-content>
@@ -62,23 +82,6 @@
                 </v-list-item>
               </template>
             </v-autocomplete>
-            <v-icon
-              v-if="dataObject.items.length > 1"
-              class="
-                control-button
-                edit-icon
-                theme--dark
-                col-2
-                pa-0
-                absolute
-                absolute--top
-                absolute--right
-              "
-              :style="{ color: theme.$secondary_text }"
-              @click="deleteLine(index, 'items')"
-            >
-              {{ closeIcon }}
-            </v-icon>
             <div
               class="col-12 mb-8 py-0 text-left"
               :style="{ color: theme.$secondary_text }"
@@ -119,11 +122,11 @@
             </v-btn>
           </div>
         </div>
-        <div class="dash-constructor-schemes__slider column align-stretch">
+        <div class="dash-cs__slider column align-stretch">
           <div class="mb-4">
             Размер подложки таблицы:
           </div>
-          <div class="dash-constructor-schemes__slider-title">
+          <div class="dash-cs__slider-title">
             <div>
               <span>{{ dataObject.widthLeft }}</span>
               <span>%</span>
@@ -140,15 +143,6 @@
             @input="updateSliderValue"
           />
         </div>
-        <v-btn
-          small
-          outlined
-          :color="theme.$primary_button"
-          :style="{ color: theme.$main_text }"
-          @click="updateModelValue"
-        >
-          Применить
-        </v-btn>
       </template>
       <template v-if="dataType === 'data-type-1'">
         <!--data-type-1-->
@@ -187,7 +181,7 @@
               <template v-slot:item="{ item, on }">
                 <v-list-item
                   ripple
-                  class="v-list-item--link dash-constructor-schemes__data-panel-select"
+                  class="v-list-item--link dash-cs__data-panel-select"
                   v-on="on"
                 >
                   <v-list-item-content>
@@ -215,15 +209,6 @@
               class="col-11 py-0"
             />
           </div>
-          <v-btn
-            small
-            outlined
-            :color="theme.$primary_button"
-            :style="{ color: theme.$main_text }"
-            @click="updateModelValue"
-          >
-            Применить
-          </v-btn>
         </div>
       </template>
       <template v-if="dataType === 'data-type-2'">
@@ -233,8 +218,28 @@
             <div
               v-for="(element, index) in dataObject.items"
               :key="`${dataObject.nodeId}-${index}`"
-              class="column dash-constructor-schemes__data-type-2"
+              class="column dash-cs__data-type-2"
             >
+              <v-btn
+                v-if="dataObject.items.length > 1"
+                ripple
+                small
+                height="20"
+                min-height="20"
+                width="20"
+                min-width="20"
+                class="dash-cs__close-btn px-0 py-0"
+                color="transparent"
+                @click="deleteLine(index, 'items')"
+              >
+                <v-icon
+                  class="theme--dark px-0 py-0"
+                  size="20"
+                  :style="{ color: theme.$main_text }"
+                >
+                  {{ closeIcon }}
+                </v-icon>
+              </v-btn>
               <div class="row align-center flex-nowrap">
                 <v-autocomplete
                   v-model="element.id"
@@ -253,7 +258,7 @@
                   <template v-slot:item="{ item, on }">
                     <v-list-item
                       ripple
-                      class="v-list-item--link dash-constructor-schemes__data-panel-select"
+                      class="v-list-item--link dash-cs__data-panel-select"
                       v-on="on"
                     >
                       <v-list-item-content>
@@ -267,14 +272,6 @@
                     </v-list-item>
                   </template>
                 </v-autocomplete>
-                <v-icon
-                  v-if="dataObject.items.length > 1"
-                  class="control-button edit-icon theme--dark col-2 pa-0"
-                  :style="{ color: theme.$secondary_text }"
-                  @click="deleteLine(index, 'items')"
-                >
-                  {{ closeIcon }}
-                </v-icon>
               </div>
               <div
                 class="mr-10 mb-3 text-left"
@@ -438,14 +435,6 @@
             </div>
           </div>
         </div>
-        <v-btn
-          small
-          outlined
-          :color="theme.$main_text"
-          @click="updateModelValue(dataObject)"
-        >
-          Применить
-        </v-btn>
       </template>
       <template v-if="dataType === 'data-type-3'">
         <!--data-type-3-->
@@ -476,7 +465,7 @@
                 <template v-slot:item="{ item, on }">
                   <v-list-item
                     ripple
-                    class="v-list-item&#45;&#45;link dash-constructor-schemes__data-panel-select"
+                    class="v-list-item&#45;&#45;link dash-cs__data-panel-select"
                     v-on="on"
                   >
                     <v-list-item-content>
@@ -494,22 +483,36 @@
             <div
               v-for="(element, index) in dataObject.imageList"
               :key="`${dataObject.nodeId}-${index}`"
-              class="column dash-constructor-schemes__data-type-3"
+              class="column dash-cs__data-type-3"
             >
+              <div class="dash-cs__close-btn">
+                <v-btn
+                  v-if="dataObject.imageList.length > 1"
+                  ripple
+                  small
+                  height="20"
+                  min-height="20"
+                  width="20"
+                  min-width="20"
+                  class="dash-cs__close-btn px-0 py-0"
+                  color="transparent"
+                  @click="deleteLine(index, 'imageList')"
+                >
+                  <v-icon
+                    class="theme--dark px-0 py-0"
+                    size="20"
+                    :style="{ color: theme.$main_text }"
+                  >
+                    {{ closeIcon }}
+                  </v-icon>
+                </v-btn>
+              </div>
               <div class="row align-start flex-nowrap">
                 <v-text-field
                   v-model="element.image"
                   label="Изображение"
                   :class="dataObject.imageList.length > 1 ? 'col-10' : 'col-10'"
                 />
-                <v-icon
-                  v-if="dataObject.imageList.length > 1"
-                  class="control-button edit-icon theme--dark col-2 px-0 py-1"
-                  :style="{ color: theme.$secondary_text }"
-                  @click="deleteLine(index, 'imageList')"
-                >
-                  {{ closeIcon }}
-                </v-icon>
               </div>
               <div class="row align-center">
                 <v-text-field
@@ -546,19 +549,11 @@
             </div>
           </div>
         </div>
-        <v-btn
-          small
-          outlined
-          :color="theme.$main_text"
-          @click="updateModelValue(dataObject)"
-        >
-          Применить
-        </v-btn>
       </template>
       <template v-if="dataType === 'data-type-4'">
         <!--data-type-4-->
         <div class="column">
-          <div class="column dash-constructor-schemes__data-type-4">
+          <div class="column dash-cs__data-type-4">
             <div class="text-left mb-2">
               Первая метрика:
             </div>
@@ -580,7 +575,7 @@
                 <template v-slot:item="{ item, on }">
                   <v-list-item
                     ripple
-                    class="v-list-item&#45;&#45;link dash-constructor-schemes__data-panel-select"
+                    class="v-list-item&#45;&#45;link dash-cs__data-panel-select"
                     v-on="on"
                   >
                     <v-list-item-content>
@@ -594,6 +589,34 @@
                   </v-list-item>
                 </template>
               </v-autocomplete>
+            </div>
+            <div
+              class="d-flex align-center align-content-center justify-space-between py-3"
+              :class="{
+                'mb-4': !dataObject.textFirstUseDefaultValue,
+              }"
+            >
+              <div class="text-left">
+                Значение по-умолчанию:
+              </div>
+              <v-switch
+                v-model="dataObject.textFirstUseDefaultValue"
+                dense
+                hide-details
+                class="ma-0"
+              />
+            </div>
+            <div
+              v-if="dataObject.textFirstUseDefaultValue"
+              class="row align-center align-content-center"
+            >
+              <div class="col-12">
+                <v-text-field
+                  v-model="dataObject.textFirstDefaultValue"
+                  dense
+                  label=""
+                />
+              </div>
             </div>
             <div class="row align-center">
               <div class="col-12">
@@ -625,9 +648,9 @@
               </div>
             </div>
           </div>
-          <div class="column dash-constructor-schemes__data-type-4">
+          <div class="column dash-cs__data-type-4">
             <div class="text-left mb-2">
-              Вторая метрика:
+              Вторая метрика(опционально):
             </div>
             <div class="row align-center mb-2">
               <v-autocomplete
@@ -647,7 +670,7 @@
                 <template v-slot:item="{ item, on }">
                   <v-list-item
                     ripple
-                    class="v-list-item&#45;&#45;link dash-constructor-schemes__data-panel-select"
+                    class="v-list-item&#45;&#45;link dash-cs__data-panel-select"
                     v-on="on"
                   >
                     <v-list-item-content>
@@ -693,8 +716,31 @@
               <div
                 v-for="(item, index) in dataObject.colors"
                 :key="`${dataObject.nodeId}-${index}`"
-                class="dash-constructor-schemes__data-type-4 dash-constructor-schemes__data-type-4--alt-bg"
+                :class="[
+                  'dash-cs__data-type-4',
+                  'dash-cs__data-type-4--alt-bg'
+                ]"
               >
+                <v-btn
+                  v-if="dataObject.colors.length > 1"
+                  ripple
+                  small
+                  height="18"
+                  min-height="18"
+                  width="18"
+                  min-width="18"
+                  class="dash-cs__close-btn px-0 py-0"
+                  color="transparent"
+                  @click="deleteLine(index, 'colors')"
+                >
+                  <v-icon
+                    class="theme--dark px-0 py-0"
+                    size="18"
+                    :style="{ color: theme.$main_text }"
+                  >
+                    {{ closeIcon }}
+                  </v-icon>
+                </v-btn>
                 <div class="row align-center justify-start align-content-start">
                   <div class="col-4">
                     <v-menu
@@ -722,7 +768,7 @@
                       />
                     </v-menu>
                   </div>
-                  <div class="col-5">
+                  <div class="col-8">
                     <v-text-field
                       v-model="item.value"
                       label="Значение"
@@ -730,33 +776,11 @@
                       class="px-0"
                     />
                   </div>
-
-                  <div class="col-3">
-                    <v-btn
-                      :disabled="dataObject.colors.length < 2"
-                      ripple
-                      small
-                      width="28"
-                      min-width="28"
-                      class="px-0 py-1"
-                      color="transparent"
-                    >
-                      <v-icon
-                        class="control-button edit-icon theme--dark pa-0"
-                        size="18px"
-                        :color="theme.$main_text"
-                        :style="{ color: theme.$main_text }"
-                        @click="deleteLine(index, 'colors')"
-                      >
-                        {{ closeIcon }}
-                      </v-icon>
-                    </v-btn>
-                  </div>
                 </div>
               </div>
             </div>
           </div>
-          <div class="column dash-constructor-schemes__data-type-4">
+          <div class="column dash-cs__data-type-4">
             <div class="text-left mb-4">
               Общие настройки:
             </div>
@@ -785,15 +809,38 @@
                 class="col-12"
               />
             </div>
+            <div class="row align-center">
+              <div class="col-8 text-left">
+                Цвет по-умолчанию:
+              </div>
+              <div class="col-4">
+                <v-menu
+                  top
+                  offset-x
+                  :close-on-content-click="false"
+                  z-index="100"
+                >
+                  <template v-slot:activator="{ on, attrs }">
+                    <v-btn
+                      :style="{
+                        'background-color': dataObject.defaultColor.rgbaString,
+                      }"
+                      dark
+                      v-bind="attrs"
+                      v-on="on"
+                    />
+                  </template>
+
+                  <v-color-picker
+                    v-model="dataObject.defaultColor.rgbaObject"
+                    dot-size="12"
+                    mode="rgba"
+                    @update:color="updateSelectedNodeColor($event, 'defaultColor')"
+                  />
+                </v-menu>
+              </div>
+            </div>
           </div>
-          <v-btn
-            small
-            outlined
-            :color="theme.$main_text"
-            @click="updateModelValue(dataObject)"
-          >
-            Применить
-          </v-btn>
         </div>
       </template>
       <template v-if="dataType === 'label-type-0'">
@@ -998,14 +1045,6 @@
             />
           </div>
         </div>
-        <v-btn
-          small
-          class="mt-4"
-          :color="theme.$primary_button"
-          @click="updateModelValue(dataObject)"
-        >
-          Применить
-        </v-btn>
       </template>
       <template v-if="dataType === 'shape-type-0'">
         <div class="row align-center text-left mb-8">
@@ -1090,14 +1129,6 @@
             />
           </div>
         </div>
-        <v-btn
-          small
-          outlined
-          :color="theme.$main_text"
-          @click="updateModelValue(dataObject)"
-        >
-          Применить
-        </v-btn>
       </template>
       <template v-if="dataType === 'edge'">
         <div class="row align-center text-left mb-8">
@@ -1149,15 +1180,17 @@
             />
           </div>
         </div>
+      </template>
+      <div class="dash-cs__data-panel-footer">
         <v-btn
           small
           outlined
           :color="theme.$main_text"
-          @click="updateModelValue(dataObject)"
+          @click="updateModelValue"
         >
           Применить
         </v-btn>
-      </template>
+      </div>
     </div>
   </div>
 </template>
@@ -1281,8 +1314,23 @@ export default {
 </script>
 
 <style lang="scss" scoped>
-.dash-constructor-schemes__data-panel-wrapper {
-  .dash-constructor-schemes__data-panel-item {
+.dash-cs__data-panel-wrapper {
+  max-height: 100%;
+  overflow-y: auto;
+  overflow-x: hidden;
+  padding-right: 10px;
+  position: relative;
+  .dash-cs__data-panel-footer {
+    background-color: var(--main_bg);
+    position: sticky;
+    bottom: 0;
+    left: 0;
+    right: 0;
+    padding-bottom: 24px;
+    padding-top: 10px;
+    z-index: 1;
+  }
+  .dash-cs__data-panel-item {
     ::v-deep.v-text-field__details {
       display: none;
     }
@@ -1297,10 +1345,11 @@ export default {
       max-width: 480px;
     }
   }
-  .dash-constructor-schemes__data-type-0,
-  .dash-constructor-schemes__data-type-2,
-  .dash-constructor-schemes__data-type-3,
-  .dash-constructor-schemes__data-type-4 {
+  .dash-cs__data-type-0,
+  .dash-cs__data-type-2,
+  .dash-cs__data-type-3,
+  .dash-cs__data-type-4 {
+    position: relative;
     border: 1px solid var(--main_border);
     border-radius: 5px;
     padding: 15px 8px;
@@ -1327,9 +1376,17 @@ export default {
         border-radius: 5px;
       }
     }
+    .dash-cs__close-btn {
+      position: absolute;
+      z-index: 1;
+      right: 2px;
+      top: 2px;
+      color: var(--main_text);
+      box-shadow: none;
+    }
   }
-  .dash-constructor-schemes__slider {
-    .dash-constructor-schemes__slider-title {
+  .dash-cs__slider {
+    .dash-cs__slider-title {
       width: inherit;
       display: flex;
       justify-content: space-between;
@@ -1337,7 +1394,7 @@ export default {
   }
 }
 
-.dash-constructor-schemes__data-panel-select {
+.dash-cs__data-panel-select {
   min-width: 220px;
   max-width: 420px;
 }