+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Constructor-schemes: The "data-type-2" element (level/storage) now has the ability to display them overlapping instead of summarizing the height of nested elements. 
+- Constructor-schemes: Added element "gate valve"(data-type-4)
 
 ### Changed